requirements.txt
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0.0
Jinja2>=3.1.0
pytest>=7.0.0
